introducing later_alloc

+++ kern/arch/i386/arch_mmu.c
@@ -91,7 +91,7 @@ static pte_t* walk_page_dir(pgindex_t *pgindex, vaddr_t *vaddr, int alloc) {
     if(*pde & PTE_P){
             pt = (pte_t*)postmap_addr(PTE_ADDR(*pde));  //PTE_ADDR + KERN_BASE
     } else {
-        if(!alloc || (pt = (pte_t *)pgalloc()) == 0)
+        if(!alloc || (pt = (pte_t *)(uint32_t)pgalloc()) == 0)
             return 0;
         memset(pt, 0, PGSIZE);  // static inline in arch-mmu.h
         *pde = premap_addr(pt) | PTE_P | PTE_W | PTE_U;
